# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Commander Companion API
  version: 0.1.0
  description: API for the Commander Companion application.
servers:
  - url: http://localhost:8080/api/v1
    description: Development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    CursorParam:
      in: query
      name: cursor
      required: false
      description: >-
        Opaque cursor returned as `next_cursor` by the previous page. It is
        returned as-is, without interpreting it. Omitting it requests the first page.
      schema:
        type: string
    LimitParam:
      in: query
      name: limit
      required: false
      description: >-
        Maximum number of items per page. Above the maximum it is silently
        clamped.
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  schemas:
...
    Deck:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        name:
          type: string
        commander:
          type: string
        moxfield_id:
          type: string
          nullable: true
        image_url:
          type: string
          nullable: true
          description: >-
            Art crop of the deck's main card (usually the commander), taken
            from Moxfield during import/resync.
...
    Game:
      type: object
      properties:
        id:
          type: string
          format: uuid
        playgroup_id:
          type: string
          format: uuid
          nullable: true
        status:
          type: string
          enum: [pending, active, finished]
        started_at:
          type: string
          format: date-time
          nullable: true
        finished_at:
          type: string
          format: date-time
          nullable: true
        current_turn_player_id:
          type: string
          format: uuid
          nullable: true
          description: Whose turn it is now (game_players.id). Null until the first TurnStart.
        players:
          type: array
          items:
            $ref: '#/components/schemas/GamePlayer'
    GameListResponse:
      type: object
      required: [items, next_cursor]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Game'
        next_cursor:
          type: string
          nullable: true
          description: >-
            Cursor to pass as the `cursor` query param to request the next
            page. It is null on the last page.
    CreateGameRequest:
      type: object
      properties:
        playgroup_id:
          type: string
          format: uuid
          description: Optional; associates the game with an existing playgroup.
    JoinGameRequest:
      type: object
      required: [deck_id]
      properties:
        deck_id:
          type: string
          format: uuid
          description: >-
            Must belong to the user who is sitting down (the authenticated
            one, or user_id if specified).
        user_id:
          type: string
          format: uuid
          description: >-
            Optional. Proxy-join (see ADR-0013): the authenticated user joins
            ANOTHER user on their behalf. Only allowed if the game has a
            playgroup_id and both are members of that same playgroup; in that
            case deck_id must belong to user_id, not the caller. If omitted
            (or equal to the authenticated user), it is a normal join — no
            change in behavior.
    GamePlayer:
      type: object
      properties:
        id:
          type: string
          format: uuid
        game_id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        deck_id:
          type: string
          format: uuid
        life_total:
          type: integer
        poison_counters:
          type: integer
        energy_counters:
          type: integer
        experience_counters:
          type: integer
        is_eliminated:
          type: boolean
    GameAction:
      type: object
      properties:
        id:
          type: string
          format: uuid
        game_id:
          type: string
          format: uuid
        actor_id:
          type: string
          format: uuid
        target_id:
          type: string
          format: uuid
          nullable: true
        action_type:
          type: string
        payload:
          type: object
        created_at:
          type: string
          format: date-time
    Playgroup:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        members:
          type: array
          items:
            $ref: '#/components/schemas/PlaygroupMember'
    PlaygroupMember:
      type: object
      properties:
        playgroup_id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        username:
          type: string
    CreatePlaygroupRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
    UpdatePlaygroupRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
    AddMemberRequest:
      type: object
      required: [user_id]
      properties:
        user_id:
          type: string
          format: uuid
          description: Must be an existing user; the one inviting must already be a member of the group.
...
    ErrorResponse:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
security:
  - bearerAuth: []
paths:
...
  /playgroups:
    get:
      summary: List the authenticated user's playgroups
      responses:
        '200':
          description: Playgroups the authenticated user is a member of
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Playgroup'
    post:
      summary: Create playgroup (the creator becomes its first member)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePlaygroupRequest'
      responses:
        '201':
          description: Playgroup created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Playgroup'
        '400':
          description: Missing playgroup name
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /playgroups/{id}:
    get:
      summary: Get playgroup
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Playgroup detail, if the authenticated user is a member
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Playgroup'
        '404':
          description: Does not exist, or the authenticated user is not a member
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    patch:
      summary: Rename the playgroup (only an existing member can edit it)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePlaygroupRequest'
      responses:
        '200':
          description: Playgroup renamed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Playgroup'
        '400':
          description: Missing new name
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Does not exist, or the authenticated user is not a member
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /playgroups/{id}/members:
    post:
      summary: Add a member to the playgroup (only an existing member can invite)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddMemberRequest'
      responses:
        '201':
          description: Member added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlaygroupMember'
        '404':
          description: Playgroup or user to add not found (or the inviter is not a member of the playgroup)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: The user is already a member of the playgroup
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /playgroups/{id}/members/{userId}/decks:
    get:
      summary: List a playgroup member's decks
      description: >-
        Allows viewing the decks of ANOTHER member of the same playgroup (see
        ADR-0013) — this is what enables choosing their deck in a proxy-join
        (`JoinGameRequest.user_id`). Authorization: both the authenticated
        user and userId must be members of this playgroup.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
        - in: path
          name: userId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Decks of the specified member
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Deck'
        '404':
          description: >-
            Playgroup does not exist, or the authenticated user is not a
            member, or userId is not a member of this playgroup (which one is
            not distinguished)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /games:
    get:
      summary: List games
      description: >-
        Returns a page of the game history, from most recent to oldest.
        Pagination is cursor-based (keyset over created_at + id): follow
        `next_cursor` until it comes back null. With `playgroup_id`, however,
        it returns the full history (unpaginated, `next_cursor` always null)
        of that playgroup — requires the authenticated user to be a member.
      parameters:
        - $ref: '#/components/parameters/CursorParam'
        - $ref: '#/components/parameters/LimitParam'
        - in: query
          name: playgroup_id
          required: false
          schema:
            type: string
            format: uuid
          description: >-
            If specified, ignores cursor/limit and returns the full history
            of the playgroup. 404 if it does not exist or the authenticated
            user is not a member.
      responses:
        '200':
          description: Page of the game history
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GameListResponse'
        '400':
          description: Invalid cursor or limit
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      summary: Create game
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGameRequest'
      responses:
        '201':
          description: Game created (pending state)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Game'
        '400':
          description: Invalid playgroup_id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /games/{id}:
    get:
      summary: Get game
      description: Includes the current state of its players.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Game detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Game'
        '404':
          description: Does not exist
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /games/{id}/join:
    post:
      summary: Join the game with your own deck, or add another player (proxy-join)
      description: >-
        Without user_id (or equal to the authenticated user): normal join, no
        changes. With a different user_id: proxy-join (see ADR-0013) — the
        caller adds another user on their behalf, only if both share the
        game's playgroup; the player ends up with added_by = caller,
        authorizing them to register actions for that seat.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JoinGameRequest'
      responses:
        '200':
          description: Player added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GamePlayer'
        '400':
          description: >-
            Invalid deck_id/user_id, or a proxy-join was attempted on a game
            without playgroup_id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: >-
            Proxy-join without shared membership: the caller or user_id do
            not belong to the game's playgroup
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Game or deck not found (or the deck does not belong to the person joining)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: The game no longer accepts players, or that user is already in the game
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /games/{id}/leave:
    post:
      summary: Leave the game (only while pending)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Player removed
        '404':
          description: The authenticated user is not a member of this game
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: The game is no longer pending
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /games/{id}/start:
    post:
      summary: Start game
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Active game
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Game'
        '409':
          description: The game is not pending, or does not have enough players
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /games/{id}/finish:
    post:
      summary: Finish game
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Game finished (triggers statistics recalculation)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Game'
        '409':
          description: Only an active game can be finished
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /games/{id}/actions:
    post:
      summary: Register an action in the game
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [actor_id, action_type]
              properties:
                actor_id:
                  type: string
                  format: uuid
                  description: ID of the GamePlayer (not the user) that originates the action.
                target_id:
                  type: string
                  format: uuid
                  description: ID of the affected GamePlayer; if omitted, it affects the actor themselves.
                action_type:
                  type: string
                  enum: [LifeChange, CombatDamage, CommanderDamage, PoisonCounter, TurnStart, TurnEnd, Elimination]
                payload:
                  type: object
                  description: 'For LifeChange/CombatDamage/CommanderDamage/PoisonCounter: { "amount": <integer> }.'
      responses:
        '400':
          description: Invalid action_type, or missing/non-numeric payload.amount
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: >-
            The authenticated user is neither the owner of the GamePlayer
            actor_id nor the one who added them as a proxy (see ADR-0013)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Game, actor, or target not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: The game is not active
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '201':
          description: Action registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GameAction'
  /games/{id}/timeline:
    get:
      summary: Get full action history
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Game timeline
...